feat: fix IModularAccount to return address of the account for native functions (#208)
The reference implementation currently returns `address(this)` which is the address of the account for `ExecutionDataView`'s `module` field. However, the standard currently contradicts this by saying we should return `address(0)` for native functions.

The better design seems to be to return the address of the account since the zero address would also be returned for nonexistent selectors, and it would be difficult to differentiate in those cases.

The client should have the address of the account handy so it'd be easy to compare it against what is returned to check if it's a native function or not.

Thanks @0xrubes for bringing this up!